+++ config/templatetags/custom_tags.py
@@ -2,8 +2,8 @@ from django.template.defaultfilters import register, slice_filter
 from django.contrib.staticfiles.finders import find
 from django.utils.six import moves
 from os.path import splitext, basename
-from base64 import b64encode
-from hashlib import sha256
+# from base64 import b64encode
+# from hashlib import sha256
 from re import sub
 
 join = moves.urllib.parse.urljoin
@@ -59,16 +59,17 @@ def get_type(link):
         }.get(ext, 'image/jpeg')
 
 
-@register.simple_tag
-def sha256hash(sfile):
-    BUF_SIZE = 4096
-    hash = sha256()
-    with open(find(sfile), 'rb') as f:
-        while True:
-            data = f.read(BUF_SIZE)
-            if not data:
-                break
-            hash.update(data)
-    b64 = b64encode(hash.digest()).decode()
-    return 'sha256-%s' % b64
+# Currently unused
+# @register.simple_tag
+# def sha256hash(sfile):
+#     BUF_SIZE = 4096
+#     hash = sha256()
+#     with open(find(sfile), 'rb') as f:
+#         while True:
+#             data = f.read(BUF_SIZE)
+#             if not data:
+#                 break
+#             hash.update(data)
+#     b64 = b64encode(hash.digest()).decode()
+#     return 'sha256-%s' % b64